+++ include/anvil/byte-pipe/BytePipeRepeater.hpp
@@ -38,16 +38,16 @@ namespace anvil { namespace BytePipe {
 
 		void OnPipeOpen() final;
 		void OnPipeClose() final;
-		void OnArrayBegin(const uint32_t size) final;
+		void OnArrayBegin(const size_t size) final;
 		void OnArrayEnd() final;
-		void OnObjectBegin(const uint32_t component_count) final;
+		void OnObjectBegin(const size_t component_count) final;
 		void OnObjectEnd() final;
 		void OnComponentID(const ComponentID id) final;
-		void OnComponentID(const char* str, const uint32_t size) final;
-		void OnUserPOD(const PodType type, const uint32_t bytes, const void* data) final;
+		void OnComponentID(const char* str, const size_t size) final;
+		void OnUserPOD(const PodType type, const size_t bytes, const void* data) final;
 		void OnNull() final;
 		void OnPrimitiveF64(const double value) final;
-		void OnPrimitiveString(const char* value, const uint32_t length) final;
+		void OnPrimitiveString(const char* value, const size_t length) final;
 		void OnPrimitiveBool(const bool value) final;
 		void OnPrimitiveC8(const char value) final;
 		void OnPrimitiveU64(const uint64_t value) final;
@@ -60,19 +60,19 @@ namespace anvil { namespace BytePipe {
 		void OnPrimitiveS16(const int16_t value) final;
 		void OnPrimitiveS32(const int32_t value) final;
 		void OnPrimitiveF16(const half value) final;
-		void OnPrimitiveArrayU8(const uint8_t* src, const uint32_t size) final;
-		void OnPrimitiveArrayU16(const uint16_t* src, const uint32_t size) final;
-		void OnPrimitiveArrayU32(const uint32_t* src, const uint32_t size) final;
-		void OnPrimitiveArrayU64(const uint64_t* src, const uint32_t size) final;
-		void OnPrimitiveArrayS8(const int8_t* src, const uint32_t size) final;
-		void OnPrimitiveArrayS16(const int16_t* src, const uint32_t size) final;
-		void OnPrimitiveArrayS32(const int32_t* src, const uint32_t size) final;
-		void OnPrimitiveArrayS64(const int64_t* src, const uint32_t size) final;
-		void OnPrimitiveArrayF32(const float* src, const uint32_t size) final;
-		void OnPrimitiveArrayF64(const double* src, const uint32_t size) final;
-		void OnPrimitiveArrayC8(const char* src, const uint32_t size) final;
-		void OnPrimitiveArrayF16(const half* src, const uint32_t size) final;
-		void OnPrimitiveArrayBool(const bool* src, const uint32_t size) final;
+		void OnPrimitiveArrayU8(const uint8_t* src, const size_t size) final;
+		void OnPrimitiveArrayU16(const uint16_t* src, const size_t size) final;
+		void OnPrimitiveArrayU32(const uint32_t* src, const size_t size) final;
+		void OnPrimitiveArrayU64(const uint64_t* src, const size_t size) final;
+		void OnPrimitiveArrayS8(const int8_t* src, const size_t size) final;
+		void OnPrimitiveArrayS16(const int16_t* src, const size_t size) final;
+		void OnPrimitiveArrayS32(const int32_t* src, const size_t size) final;
+		void OnPrimitiveArrayS64(const int64_t* src, const size_t size) final;
+		void OnPrimitiveArrayF32(const float* src, const size_t size) final;
+		void OnPrimitiveArrayF64(const double* src, const size_t size) final;
+		void OnPrimitiveArrayC8(const char* src, const size_t size) final;
+		void OnPrimitiveArrayF16(const half* src, const size_t size) final;
+		void OnPrimitiveArrayBool(const bool* src, const size_t size) final;
 	};
 
 }}
